#231: Add `dry_run` and optional `overwrite` to sampling workflow

+++ romtools/workflows/sampling/sampling.py
@@ -43,6 +43,7 @@
 # ************************************************************************
 #
 
+import os
 import time
 import numpy as np
 
@@ -59,7 +60,9 @@ def run_sampling(model: Model,
                  parameter_space: ParameterSpace,
                  absolute_sampling_directory: str,
                  number_of_samples: int = 10,
-                 random_seed: int = 1):
+                 random_seed: int = 1,
+                 dry_run: bool = False,
+                 overwrite: bool = False):
     '''
     Core algorithm
     '''
@@ -70,7 +73,7 @@ def run_sampling(model: Model,
     parameter_samples = parameter_space.generate_samples(number_of_samples)
     parameter_names = parameter_space.get_names()
 
-    # Setup model directories
+    # Set up model directories
     run_directory_base = f'{absolute_sampling_directory}/run_'
     run_directories = []
     starting_sample_index = 0
@@ -82,18 +85,21 @@ def run_sampling(model: Model,
         model.populate_run_directory(run_directory, parameter_dict)
         run_directories.append(run_directory)
 
-    # Run cases
-    run_times = np.zeros(number_of_samples)
-    for sample_index in range(0, number_of_samples):
-        print("=======  Sample " + str(sample_index) + " ============")
-        print("Running")
-        run_directory = f'{run_directory_base}{sample_index}'
-        parameter_dict = _create_parameter_dict(parameter_names, parameter_samples[sample_index])
-        run_times[sample_index] = run_sample(run_directory, model,
-                                             parameter_dict)
-        sample_stats_save_directory = f'{run_directory_base}{sample_index}/../'
-        np.savez(f'{sample_stats_save_directory}/sampling_stats',
-                 run_times=run_times)
+    # Run cases if dry_run is not set
+    if not dry_run:
+        run_times = np.zeros(number_of_samples)
+        for sample_index in range(0, number_of_samples):
+            print("=======  Sample " + str(sample_index) + " ============")
+            run_directory = f'{run_directory_base}{sample_index}'
+            if "passed.txt" in os.listdir(run_directory) and not overwrite:
+                print("Skipping (Sample has already run successfully)")
+            else:
+                print("Running")
+                parameter_dict = _create_parameter_dict(parameter_names, parameter_samples[sample_index])
+                run_times[sample_index] = run_sample(run_directory, model, parameter_dict)
+                sample_stats_save_directory = f'{run_directory_base}{sample_index}/../'
+                np.savez(f'{sample_stats_save_directory}/sampling_stats',
+                            run_times=run_times)
 
     return run_directories
 
@@ -110,6 +116,7 @@ def run_sample(run_directory: str, model: Model,
     run_time = tf - ts
 
     if flag == 0:
+        np.savetxt(os.path.join(run_directory, 'passed.txt'), np.array([0]), '%i')
         print(f"Sample complete, run time = {run_time}")
     else:
         print(f"Sample failed, run time = {run_time}")
